% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiece.R
\name{HourlyPieceServer_detail_view}
\alias{HourlyPieceServer_detail_view}
\title{按明细查询}
\usage{
HourlyPieceServer_detail_view(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
按明细查询
}
\examples{
HourlyPieceServer_detail_view()
}
