% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiece.R
\name{HourlyPieceServer_download}
\alias{HourlyPieceServer_download}
\title{按下载数据}
\usage{
HourlyPieceServer_download(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
按下载数据
}
\examples{
HourlyPieceServer_download()
}
