% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyPiece.R
\name{HourlyPieceServer_preview}
\alias{HourlyPieceServer_preview}
\title{预览}
\usage{
HourlyPieceServer_preview(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
预览
}
\examples{
HourlyPieceServer_preview()
}
