% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyRouting.R
\name{HourlyRoutingServer_delete}
\alias{HourlyRoutingServer_delete}
\title{按单据编号和行号删除}
\usage{
HourlyRoutingServer_delete(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
按单据编号和行号删除
}
\examples{
HourlyRoutingServer_delete()
}
