% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HourlyWages.R
\name{HourlyWagesServer_view}
\alias{HourlyWagesServer_view}
\title{按工序名称查询}
\usage{
HourlyWagesServer_view(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
按工序名称查询
}
\examples{
HourlyWagesServer_view()
}
