% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productdaily.R
\name{productdailyServer_erpsync}
\alias{productdailyServer_erpsync}
\title{同步数据}
\usage{
productdailyServer_erpsync(input, output, session, dms_token, erp_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
同步数据
}
\examples{
productdailyServer_erpsync()
}
