% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productdaily.R
\name{productdailyServer_get}
\alias{productdailyServer_get}
\title{获取工序生产日报}
\usage{
productdailyServer_get(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
获取工序生产日报
}
\examples{
productdailyServer_get()
}
