% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productprice.R
\name{productpriceServer_get}
\alias{productpriceServer_get}
\title{获取订单核价单}
\usage{
productpriceServer_get(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
获取订单核价单
}
\examples{
productpriceServer_get()
}
