% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productRouting.R
\name{productRoutingServer_upload}
\alias{productRoutingServer_upload}
\title{上传}
\usage{
productRoutingServer_upload(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
上传
}
\examples{
productRoutingServer_upload()
}
